Initial commit of completed tools
Each tool includes a graphical interface with instructions on how to run the tool and portals for user input.

The tools focus on manipulating an object (or a group of objects) in reference to other objects in different ways. All objects are selected by the user, and can be "any" objects that follow the standard naming convention for copied objects in Maya. It is assumed CV's or UV's will be used for the object references, and that a polygonal or NURBS surface will be used for the manipulated object. The scripts therefore may not work with all objects.

For more details, run any of these tools and follow the provided instructions.